+++ server/src/test/java/org/realityforge/replicant/server/ee/rest/SessionRestServiceTest.java
@@ -2,7 +2,7 @@ package org.realityforge.replicant.server.ee.rest;
 
 import javax.ws.rs.core.Response;
 import org.realityforge.replicant.server.TestSession;
-import org.realityforge.ssf.SessionManager;
+import org.realityforge.replicant.server.transport.ReplicantSessionManager;
 import org.testng.annotations.Test;
 import static org.mockito.Mockito.*;
 import static org.testng.Assert.*;
@@ -13,7 +13,7 @@ public class SessionRestServiceTest
   public void poll_dataAvailable()
     throws Exception
   {
-    final SessionManager sessionManager = mock( SessionManager.class );
+    final ReplicantSessionManager sessionManager = mock( ReplicantSessionManager.class );
     final AbstractSessionRestService resource = newResource( sessionManager );
 
     when( sessionManager.createSession() ).
@@ -24,13 +24,13 @@ public class SessionRestServiceTest
     assertEquals( token.getEntity(), "2222" );
   }
 
-  protected AbstractSessionRestService newResource( final SessionManager sessionManager )
+  protected AbstractSessionRestService newResource( final ReplicantSessionManager sessionManager )
     throws Exception
   {
     return new AbstractSessionRestService()
     {
       @Override
-      protected SessionManager getSessionManager()
+      protected ReplicantSessionManager getSessionManager()
       {
         return sessionManager;
       }